# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with authentication, roles and encryption.",
        "title": "Go API Dojo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/v1/categories": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List all course categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CourseCategory"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.CourseCategory"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get a category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CourseCategory"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "category not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CourseCategory"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "category not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fault.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "a descriptive error message"
                }
            }
        },
...
        "handler.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "color_hex": {
                    "type": "string",
                    "maxLength": 7
                },
                "description": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string",
                    "maxLength": 512
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "parent_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateCategoryRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "color_hex": {
                    "type": "string",
                    "maxLength": 7
                },
                "description": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string",
                    "maxLength": 512
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "parent_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "model.CourseCategory": {
            "type": "object",
            "properties": {
                "colorHex": {
                    "type": "string"
                },
                "courseCount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "iconURL": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isVisible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parentID": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}